Full Budget/Tax Macro
This VBA code will allow you to estimate taxes and create a budget in Microsoft Excel. 
Note: Requires the city and state that you work in to be in Cells C2 and A2, respectively, and the state and city that you live in to be in cells F2 and H2, respectively.  Then run entire macro from the top.

It is also possible to re-estimate taxes once the macro has been run.  To estimate your taxes with different income, location/s and/or pre-tax deductions, change the desired outcome in the spreadsheet and then run the macro "Re-Estimate Taxes." To update anything in your budget, simply change it and the spreadsheet will automatically update.